feat(animation): add CAnimation and CAnimator classes for managing animations
- Added CAnimation class to handle individual animations:
  - Includes frame information (tAnimFrm)
  - Manages current frame, accumulated time, and animation state

- Added CAnimator class to manage multiple animations:
  - Stores animations in a map
  - Tracks current animation and owner object

- Integrated CAnimator for texture loading in CObject class
  - Updated texture management to use CAnimator

+++ Win32API_Introduction/CPathMgr.cpp
@@ -14,7 +14,7 @@ CPathMgr::~CPathMgr()
 void CPathMgr::Init()
 {
 	GetCurrentDirectory(255, m_szContentPath);
-	int iLength = wcslen(m_szContentPath);
+	int iLength = (int)wcslen(m_szContentPath);
 
 	// Parent Path
 	for (int i = iLength - 1; i >= 0; --i)
